namespace issue with ruby load error - fixed
