+++ alfresco4r.gemspec
@@ -12,7 +12,7 @@ Gem::Specification.new do |gem|
   gem.summary       = %q{Interaction with Alfresco from Ruby}
   gem.homepage      = "https://github.com/sivaprakasamboopathy/alfresco4r"
 
-  gem.files         =  Dir.glob("lib/*/*.rb")
+  gem.files         =  Dir.glob("lib/**/*.rb")
   gem.executables   = gem.files.grep(%r{^bin/}).map{ |f| File.basename(f) }
   gem.test_files    = gem.files.grep(%r{^(test|spec|features)/})
   gem.name          = "alfresco4r"
