+++ lib/alfresco4r/version.rb
@@ -1,3 +1,3 @@
 module Alfresco4r
-  VERSION = "0.0.2"
+  VERSION = "1.0.0"
 end
